+++ packages/react/Badge/src/badge.types.ts
@@ -5,9 +5,11 @@ export interface BadgeComponents {
   Skeleton: typeof Skeleton;
 }
 
-export interface BadgeProps extends HTMLAttributes<HTMLHeadingElement> {
+export interface BadgeProps extends HTMLAttributes<HTMLElement> {
   /** Badge color */
   appearance?: "primary" | "success" | "warning" | "danger" | "neutral";
   /** Total items to be displayed without badge */
   count: number | string;
+  /** Badge appearance color scheme */
+  theme?: "surface" | "light";
 }