+++ packages/react/Chip/src/chip.types.ts
@@ -1,11 +1,11 @@
-import { HTMLAttributes, ReactNode } from "react";
+import { ReactNode, ButtonHTMLAttributes } from "react";
 import { Skeleton } from "./components";
 
 export interface ChipComponents {
   Skeleton: typeof Skeleton;
 }
 
-export interface ChipProps extends HTMLAttributes<HTMLButtonElement> {
+export interface ChipProps extends ButtonHTMLAttributes<HTMLButtonElement> {
   /** SVG icon to be displayed on chip */
   icon?: ReactNode;
   /** Informs whether the close icon should be displayed */