+++ packages/react/Toggle/src/index.ts
@@ -1,0 +1,6 @@
+import { Toggle } from "./Toggle";
+import "@nimbus-ds/styles/packages/toggle/index.css";
+
+export { Toggle } from "./Toggle";
+export type { ToggleProps } from "./toggle.types";
+export default Toggle;