+++ packages/react/Badge/CHANGELOG.md
@@ -1,6 +1,26 @@
 # Changelog
 
-Badges are used to inform merchants of the status of an object or of an action that’s been taken.
+The badge component is used to communicate quantities of items or pending actions.
+
+## 2022-09-29
+
+### 🎉 New features
+
+- Added displayName to component `Badge.Skeleton`. ([#32](https://github.com/TiendaNube/nimbus-design-system/pull/32) by [@juniorconquista](https://github.com/juniorconquista))
+
+## 2022-09-28
+
+### 🎉 New features
+
+- Added `theme` properties to the component. ([#32](https://github.com/TiendaNube/nimbus-design-system/pull/32) by [@juniorconquista](https://github.com/juniorconquista))
+- Added displayName to component. ([#32](https://github.com/TiendaNube/nimbus-design-system/pull/32) by [@juniorconquista](https://github.com/juniorconquista))
+- Added displayName to component `Badge.Skeleton`. ([#32](https://github.com/TiendaNube/nimbus-design-system/pull/32) by [@juniorconquista](https://github.com/juniorconquista))
+
+## 2022-09-27
+
+### 💡 Others
+
+- Adjusted component native typing for HTMLElement. ([#30](https://github.com/TiendaNube/nimbus-design-system/pull/30) by [@juniorconquista](https://github.com/juniorconquista))
 
 ## 2022-09-01