+++ packages/react/Checkbox/CHANGELOG.md
@@ -2,6 +2,23 @@
 
 The Checkbox component allows us to choose one or several options from a selection list.
 
+## 2022-09-29
+
+### 🎉 New features
+
+- Added displayName to component. ([#32](https://github.com/TiendaNube/nimbus-design-system/pull/32) by [@juniorconquista](https://github.com/juniorconquista))
+- Added displayName to component `Checkbox.Skeleton`. ([#32](https://github.com/TiendaNube/nimbus-design-system/pull/32) by [@juniorconquista](https://github.com/juniorconquista))
+
+### 💡 Others
+
+- Adjusted color Icon check from `primary.background` to `neutral.background`. ([#32](https://github.com/TiendaNube/nimbus-design-system/pull/32) by [@juniorconquista](https://github.com/juniorconquista))
+
+## 2022-09-27
+
+### 💡 Others
+
+- Adjusted component native typing for InputHTMLAttributes. ([#30](https://github.com/TiendaNube/nimbus-design-system/pull/30) by [@juniorconquista](https://github.com/juniorconquista))
+
 ## 2022-09-24
 
 ### 💡 Others